+++ dloserver/test/unit/com/dlohaiti/dloserver/ProductBuilder.groovy
@@ -3,7 +3,7 @@ package com.dlohaiti.dloserver
 class ProductBuilder {
   String sku = "ABC"
   String description = "A thing for sale"
-  Integer gallons = 10
+  Double gallons = 10.0
   Money price = new Money(amount: BigDecimal.ONE)
   String reportingCategory = "FYO"
   Integer minimumQuantity
