+++ dloserver/test/unit/com/dlohaiti/dloserver/ReceiptLineItemBuilder.groovy
@@ -6,7 +6,7 @@ class ReceiptLineItemBuilder {
   String type = "PRODUCT"
   BigDecimal price = 10G
   String currencyCode = "HTG"
-  Integer gallons = 10
+  Double gallons = 10.0
 
   ReceiptLineItem build() {
     return new ReceiptLineItem(sku: sku, quantity: quantity, type: type, price: price, currencyCode: currencyCode, gallons: gallons)
